requirements.txt
mysql-connector-python==8.0.20
psycopg2==2.8.5
pandas==1.0.4
SQLAlchemy~=1.3.17